Refactored game networking to make more sense
