--- core/src/edu/rice/rbox/Game/Common/Utils/GameServerGrpc.java
@@ -1,43 +1,0 @@
-package edu.rice.rbox.Game.Common.Utils;
-
-import io.grpc.Server;
-import io.grpc.ServerBuilder;
-import java.io.IOException;
-import java.net.DatagramSocket;
-import java.net.InetAddress;
-import java.net.SocketException;
-import java.net.UnknownHostException;
-
-public class GameServerGrpc {
-
-  public static void main(String args[]) throws IOException, InterruptedException {
-    String ip = "";
-    try(final DatagramSocket socket = new DatagramSocket()){
-      socket.connect(InetAddress.getByName("8.8.8.8"), 10002);
-      ip = socket.getLocalAddress().getHostAddress();
-    } catch (SocketException | UnknownHostException e) {
-      System.err.println("Server failed to start.");
-      e.printStackTrace();
-    }
-
-    System.out.println("Server Running on address: " + ip);
-
-
-    // Create a new server to listen on port 8080
-
-    Server server = ServerBuilder.forPort(8080)
-                        .addService(new GameGrpc())
-                        .build();
-
-
-    // Start the server
-    server.start();
-
-    // Server threads are running in the background.
-    System.out.println("Server started");
-
-    // Don't exit the main thread. Wait until server is terminated.
-    server.awaitTermination();
-  }
-
-}
