--- core/src/edu/rice/rbox/Game/Server/ClientConnectionManager.java
@@ -1,7 +1,0 @@
-package edu.rice.rbox.Game.Server;
-
-public class ClientConnectionManager {
-
-
-
-}
